Move schema and type definitions to repository layer
- Move Command schema and type definitions from models to repository layer
- Remove redundant CommandHistory model file
- Simplify imports in Shell component
